more printf fixes

+++ src/cpu/core_normal/prefix_0f.h
@@ -156,7 +156,7 @@ bool CPU_WRMSR();
 			Bitu which=(rm >> 3) & 7;
 			if (rm < 0xc0 ) {
 				rm |= 0xc0;
-				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,CR%u with non-register",which);
+				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,CR%u with non-register",(int)which);
 			}
 			GetEArd;
 			Bit32u crx_value;
@@ -170,7 +170,7 @@ bool CPU_WRMSR();
 			Bitu which=(rm >> 3) & 7;
 			if (rm < 0xc0 ) {
 				rm |= 0xc0;
-				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,DR%u with non-register",which);
+				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,DR%u with non-register",(int)which);
 			}
 			GetEArd;
 			Bit32u drx_value;
@@ -184,7 +184,7 @@ bool CPU_WRMSR();
 			Bitu which=(rm >> 3) & 7;
 			if (rm < 0xc0 ) {
 				rm |= 0xc0;
-				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,CR%u with non-register",which);
+				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,CR%u with non-register",(int)which);
 			}
 			GetEArd;
 			if (CPU_WRITE_CRX(which,*eard)) RUNEXCEPTION();
@@ -196,7 +196,7 @@ bool CPU_WRMSR();
 			Bitu which=(rm >> 3) & 7;
 			if (rm < 0xc0 ) {
 				rm |= 0xc0;
-				LOG(LOG_CPU,LOG_ERROR)("MOV DR%u,XXX with non-register",which);
+				LOG(LOG_CPU,LOG_ERROR)("MOV DR%u,XXX with non-register",(int)which);
 			}
 			GetEArd;
 			if (CPU_WRITE_DRX(which,*eard)) RUNEXCEPTION();
@@ -208,7 +208,7 @@ bool CPU_WRMSR();
 			Bitu which=(rm >> 3) & 7;
 			if (rm < 0xc0 ) {
 				rm |= 0xc0;
-				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,TR%u with non-register",which);
+				LOG(LOG_CPU,LOG_ERROR)("MOV XXX,TR%u with non-register",(int)which);
 			}
 			GetEArd;
 			Bit32u trx_value;
@@ -222,7 +222,7 @@ bool CPU_WRMSR();
 			Bitu which=(rm >> 3) & 7;
 			if (rm < 0xc0 ) {
 				rm |= 0xc0;
-				LOG(LOG_CPU,LOG_ERROR)("MOV TR%u,XXX with non-register",which);
+				LOG(LOG_CPU,LOG_ERROR)("MOV TR%u,XXX with non-register",(int)which);
 			}
 			GetEArd;
 			if (CPU_WRITE_TRX(which,*eard)) RUNEXCEPTION();
